+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.10.0 (TBD)
 
+### Enhancements
+
+- Added support for gRPC server side reflection to all components (#949).
+
 ## v0.9.0 (2025-05-30)
 
 ### Enhancements